requirements.txt
blinker==1.9.0
click
Flask==2.3.3
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
six==1.17.0
Werkzeug==3.1.3
WTForms==3.2.1
pikepdf>=8.0.0
anthropic>=0.39.0
Pillow>=10.0.0
